Split TaskConfig off of App
This makes TaskConfig a many-to-one with App, allowing multiple
different providers to be configured for an App. This allows different
configurations for deploy/build type tasks.

Right now, this is fully backwards compatible and assumes that the only
type of TaskConfig that will exist is "deploy". This allows the API to
be expanded next to add build configs.