add a datadog notifier to freight
It would be handy to be able to visually see when software is deployed in the datadog ui.  This notifier, when combined with the requisite freight-cli changes to add the notifier's webhook_url, should display a vertical bar on dd graphs whenever a deployment happens.

Datadog Reference https://docs.datadoghq.com/api/?lang=bash#post-an-event