add a datadog notifier to freight (#155)
* add a datadog notifier to freight

It would be handy to be able to visually see when software is deployed in the datadog ui.  This notifier, when combined with the requisite freight-cli changes to add the notifier's webhook_url, should display a vertical bar on dd graphs whenever a deployment happens.

Datadog Reference https://docs.datadoghq.com/api/?lang=bash#post-an-event

* update tests to include datadog notifiers

* fix lint errors

* add test_datadog.py

travis is failing because it can't find this test stub to run.

* make the webhook payload more meaningful

added in comments around where to find datadog event details as well as how the payload is parsed.
Changed the webhook_url to api_webhook_url to make it more unique, per Alex's suggestion

* clean up copy pasta

I missed several "Sentry" entries that should now be "Datadog", which are causing build tests to fail.

* reverting api_webhook_url name change

The variable is consistently used as webhook_url throughout all of the tests, so let's remain consistent

* try without slack test to confirm dd test works

I think that the test doesn't like having multiple notifiers defined, so I'm going to test datadog on it's own, and if that works, then I'll have confidence in the main part of the change and then it's a matter of getting the tests to do the whole thing.

* make tags/release info more readible

also revert my datadog testing changes.

* remove extranous datadog tags

Tags are universsal, so add more cardinality to the individual tags and only stuff useful things into dd

* moved event_title to a function

this is used between slack and datadog notifiers so it should be accessible by both of them

also fixed up python issues

* include event where necessary

* include the correct things to make python happy

* remove unused includes

* can't concatinate str + int, so convert the int

* fixing lint errors

* make freight test the new notifier configs

upon attempting to add the notifier config to freight via freight-cli, I learned that it tests the config before accepting it, via these files.  I think that this should allow for the testing to do the right thing now.