provider: Add pipeline provider (#197)
This change provides a new pipeline provider. A pipeline provider is comprised of steps. The 3 steps being implemented currently are `Shell`, `KubernetesJob`, and `KubernetesDeployment`.

Each step is run sequentially, and if applicable, within a step, in parallel. If any step fails, the rest of the pipeline does not continue.

## Kubernetes

There is a top level, optional, `kubernetes` configuration block. This controls how to fetch kubernetes credentials, or which context to use for authentication. One of either `context` or `credentials` may be defined, but not both. At the moment, `credentials` only supports sourcing from `gcloud`.

## Steps

### Shell

This mirrors behavior of our current `shell` provider, just as a step. No new behavior.

### KubernetesJob

This provides the ability to run a one-off `Job` spec in Kubernetes. The struct here is pared down a bit to not support the full breadth of a `Job` in Kubernetes, but just a subset in order to provide more control.

### KubernetesDeploy

This effectively is running `kubectl set image ...` and monitors the rollout of this change. You can target multiple `Deployments` in Kubernetes by the `selector` block, and `containers` block.

## Example

```
"provider": "pipeline",
"provider_config": {
    "steps": [
      {
        "kind": "Shell", 
        "command": "echo ref={ref} environment={environment}"
      }, 
      {
        "kind": "KubernetesJob", 
        "args": [
          "bash", 
          "-c", 
          "for i in {{1..30}}; do echo hello $i && sleep 0.5; done"
        ], 
        "name": "upgrade", 
        "image": "debian"
      }, 
      {
        "kind": "KubernetesDeployment", 
        "containers": [
          {
            "image": "sentry:{sha}", 
            "name": "sentry"
          }
        ], 
        "selector": {
          "label_selector": "app=sentry,tier=some-tier,environment={environment}"
        }
      }
    ], 
    "kubernetes": {
      "credentials": {
        "kind": "gcloud", 
        "project": "mattrobenolt-xxx", 
        "cluster": "xxx", 
        "zone": "us-west1-c"
      }
    }
  }, 
```