feat(remotevcs): Add /remote-changes/ API
This API returns a list of changes from the remote, given a start and
end ref.

Currently this is implemented for the GitHub VCS remote, and will result
in a list of pull commits with associatedPullRequests and externalCommit
keys, which will be useful for the UI to generate a concise list of
changes.